+++ spring-batch-core/src/main/java/org/springframework/batch/core/step/skip/SkipLimitExceededException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2007 the original author or authors.
+ * Copyright 2006-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,18 +23,19 @@ package org.springframework.batch.core.step.skip;
  * @author Ben Hale
  * @author Lucas Ward
  * @author Dave Syer
+ * @author Mahmoud Ben Hassine
  */
 @SuppressWarnings("serial")
 public class SkipLimitExceededException extends SkipException {
 
-	private final int skipLimit;
+	private final long skipLimit;
 	
-	public SkipLimitExceededException(int skipLimit, Throwable t) {
+	public SkipLimitExceededException(long skipLimit, Throwable t) {
 		super("Skip limit of '" + skipLimit + "' exceeded", t);
 		this.skipLimit = skipLimit;
 	}
 	
-	public int getSkipLimit() {
+	public long getSkipLimit() {
 	    return skipLimit;
     }
 }
